Rename test file to internal
